use operand for fetch-add

+++ mooncake-transfer-engine/flex_transfer_engine/flex_batch.h
@@ -40,7 +40,7 @@ class FlexBatch {
     void addWriteRequest(uintptr_t local_addr, uintptr_t remote_addr,
                          uint64_t size);
     void addFetchAddRequest(uintptr_t local_addr, uintptr_t remote_addr,
-                            uint64_t value);
+                            int32_t value);
 
     /**
      * Submit the transfer batch.
